Adds is_prerelease to NVR
